in_progress/cloud_run/02_-_image_processing_app/requirements-test.txt
pytest==6.2.4
mock==4.0.3
google-cloud-pubsub==2.9.0